changed to doubles

+++ src/deconv/UB_spe.h
@@ -15,8 +15,8 @@ namespace wcopreco {
      UB_spe(std::string nm, bool mult_flag);
      virtual ~UB_spe() {};
 
-    std::vector<float> Get_wfm(int nbins, float tick_width_ns);
-    std::vector<float> Get_pow_spec(int nbins, float tick_width_ns);
+    std::vector<double> Get_wfm(int nbins, float tick_width_ns);
+    std::vector<double> Get_pow_spec(int nbins, float tick_width_ns);
     float gain;
 
   protected:
